// src/types/StationLinks.ts
/**************************************************
 * PetrolScan Crawler
 * @author Marek Poláček (POL0423)
 * @version 0.0.1
 * @description Web crawler for my Bachelor Thesis assignment: Fuel Price Comparison App
 * @license MIT
 * @link https://github.com/pol0423/petrolscan-crawler
 * 
 * @see https://crawlee.dev
 * 
 * File: src/types/StationLinks.ts
 */

// Imports
//-------------------------------------------------

// No imports

// Logic
//-------------------------------------------------

type StationLinks = {
    location: string;
    href: string;
}

export default StationLinks;